ClaudeCodeによるディレクトリ修正
