+++ src/App.tsx
@@ -1,24 +1,27 @@
 // App.tsx
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import HomePage from "./HomePage";
-import CalendarPage from "./CalendarPage";
-import RecipeList from "./RecipeList";
-import RecipeDetail from "./RecipeDetail";
-import RecipeListPage from "./RecipeListPage";
-import RecipeItemDetail from "./RecipeItemDetail";
+import LandingPage from "./pages/LandingPage";
+import CalendarPage from "./pages/CalendarPage";
+import DailyRecipesPage from "./components/recipe/DailyRecipesPage";
+import DailyRecipeDetailPage from "./components/recipe/DailyRecipeDetailPage";
+import AllRecipesPage from "./pages/AllRecipesPage";
+import RecipeDetailPage from "./components/recipe/RecipeDetailPage";
 
 const App: React.FC = () => {
   return (
     <Router>
       <div className="min-h-screen bg-gray-100">
         <Routes>
-          <Route path="/" element={<HomePage />} />
+          <Route path="/" element={<LandingPage />} />
           <Route path="/calendar" element={<CalendarPage />} />
-          <Route path="/recipes" element={<RecipeListPage />} />
-          <Route path="/recipes/detail" element={<RecipeItemDetail />} />
-          <Route path="/recipe/:date" element={<RecipeList />} />
-          <Route path="/recipe/:date/detail" element={<RecipeDetail />} />
+          <Route path="/recipes" element={<AllRecipesPage />} />
+          <Route path="/recipes/detail" element={<RecipeDetailPage />} />
+          <Route path="/recipe/:date" element={<DailyRecipesPage />} />
+          <Route
+            path="/recipe/:date/detail"
+            element={<DailyRecipeDetailPage />}
+          />
         </Routes>
       </div>
     </Router>
